server/requirements.txt
django
djangorestframework
django-environ
django-cors-headers
requests